Add support for Chrome OS
This commit declares that using the touch screen is optional. This is believed to be what's necessary to add support to Chrome OS, however this stands untested.

This also updates the ImagePicker library, as that will turn the Camera permission optional. With the camera permission being optional, more chromebooks will be able to use the app, as they don't have cameras.

I couldn't execute the emulator in two different PCs, and I don't own a chromebook, so this will have to be field-tested.

Closes #39